BaseDatabase Class

+++ app/src/main/java/com/hihasan/constant/DatabaseConstant.kt
@@ -2,7 +2,7 @@ package com.hihasan.constant
 
 object DatabaseConstant {
     var database_name : String = "notes.db"
-    var database_version : Long = 1
+    var database_version : Int = 1
 
     var TBL_NOTE : String = "tbl_note"
 }
